+++ data/src/main/java/hera/database/entities/Localisation.java
@@ -12,7 +12,7 @@ public class Localisation implements PersistenceEntity {
 
 	@Id
 	@GeneratedValue(strategy = GenerationType.IDENTITY)
-	private int id;
+	private Long id;
 
 	private String language;
 
@@ -31,18 +31,18 @@ public class Localisation implements PersistenceEntity {
 		this.value = value;
 	}
 
-	public Localisation(int id, String language, LocalisationKey key, String value) {
+	public Localisation(Long id, String language, LocalisationKey key, String value) {
 		this.id = id;
 		this.language = language;
 		this.key = key;
 		this.value = value;
 	}
 
-	public int getId() {
+	public Long getId() {
 		return id;
 	}
 
-	public void setId(int id) {
+	public void setId(Long id) {
 		this.id = id;
 	}
