+++ environment/sql/tables/default_role.sql
@@ -1,5 +1,5 @@
 CREATE TABLE `default_role` (
-  `id` int(11) PRIMARY KEY AUTO_INCREMENT,
-  `guildFK` int(11) NOT NULL,
-  `roleFK` int(11) NOT NULL
+  `id` bigint(11) PRIMARY KEY AUTO_INCREMENT,
+  `guildFK` bigint(11) NOT NULL,
+  `roleFK` bigint(11) NOT NULL
 ) ENGINE=InnoDB DEFAULT CHARSET=latin1;
