+++ environment/sql/tables/global_setting.sql
@@ -1,5 +1,5 @@
 CREATE TABLE `global_setting` (
-  `id` int(11) PRIMARY KEY AUTO_INCREMENT,
+  `id` bigint(11) PRIMARY KEY AUTO_INCREMENT,
   `name` varchar(50) NOT NULL,
   `value` varchar(500) NOT NULL
 ) ENGINE=InnoDB DEFAULT CHARSET=latin1;
