+++ environment/sql/tables/localisation.sql
@@ -1,5 +1,5 @@
 CREATE TABLE `localisation` (
-  `id` int(11) PRIMARY KEY AUTO_INCREMENT,
+  `id` bigint(11) PRIMARY KEY AUTO_INCREMENT,
   `language` varchar(10) NOT NULL,
   `name` varchar(50) NOT NULL,
   `value` varchar(1000) NOT NULL
