+++ environment/sql/tables/snowflake_type.sql
@@ -1,4 +1,4 @@
 CREATE TABLE `snowflake_type` (
-  `id` int(11) PRIMARY KEY AUTO_INCREMENT,
+  `id` bigint(11) PRIMARY KEY AUTO_INCREMENT,
   `type` varchar(30) NOT NULL
 ) ENGINE=InnoDB DEFAULT CHARSET=latin1;
